add read only json file in release

+++ app/src/main/java/com/a5k/tasksheduler/data/converter/Converter.kt
@@ -3,6 +3,6 @@ package com.a5k.tasksheduler.data.converter
 import com.a5k.tasksheduler.data.model.TaskDto
 import com.a5k.tasksheduler.domain.entity.Task
 
-fun TaskDto.toTask() = Task(id, dateStart.toLong(), dateFinish.toLong(), name, description)
+fun TaskDto.toTask() = Task(id, dateStart, dateFinish, name, description)
 
 fun Task.toTaskDto() = TaskDto(id, dateStart, dateFinish, name, description)
